1. 新增技能 DashSlash 2. 技能效果、架構重修 3. 技能欄位顯示錯誤修正 4. 新增烈焰噴射技能提示 5. 操作修正 (有時動作會被暈眩等異常狀態卡住) 6. AI權重邏輯修正 (現在會更符合預期動作) 7. 玩家傷害顯示修正

+++ 掠食者/Assets/Script/Character/Skills/Instances/SelfHeal.cs
@@ -10,6 +10,6 @@
     ///</summary>
     private void Heal()
     {
-        sourceCaster.CurrentHealth += sourceCaster.data.maxHealth.Value * 0.1f + sourceCaster.data.status.dexterity.Value * 5 + sourceCaster.data.status.intelligence.Value * 10;
+        sourceCaster.CurrentHealth += (int)(sourceCaster.data.maxHealth.Value * 0.1f + sourceCaster.data.status.dexterity.Value * 5 + sourceCaster.data.status.intelligence.Value * 10);
     }
 }
